XRDDEV-801: Remove access rights linked to endpoint on endpoint removal

+++ src/proxy-ui-api/src/main/java/org/niis/xroad/restapi/repository/EndpointRepository.java
@@ -24,15 +24,18 @@
  */
 package org.niis.xroad.restapi.repository;
 
+import ee.ria.xroad.common.conf.serverconf.dao.ClientDAOImpl;
+import ee.ria.xroad.common.conf.serverconf.model.ClientType;
 import ee.ria.xroad.common.conf.serverconf.model.EndpointType;
 
 import lombok.extern.slf4j.Slf4j;
 import org.hibernate.Session;
+import org.niis.xroad.restapi.service.ClientNotFoundException;
 import org.niis.xroad.restapi.service.EndpointService;
+import org.niis.xroad.restapi.service.NotFoundException;
 import org.niis.xroad.restapi.util.PersistenceUtils;
 import org.springframework.stereotype.Repository;
 import org.springframework.transaction.annotation.Transactional;
-
 
 @Slf4j
 @Repository
@@ -61,15 +64,27 @@ public class EndpointRepository {
      * @param id
      * @throws EndpointService.EndpointNotFoundException
      */
-    public void delete(String id) throws EndpointService.EndpointNotFoundException {
+    public void delete(String id) throws ClientNotFoundException, EndpointService.EndpointNotFoundException {
         Session session = this.persistenceUtils.getCurrentSession();
         EndpointType endpointType = session.get(EndpointType.class, Long.valueOf(id));
 
         if (endpointType == null) {
             throw new EndpointService.EndpointNotFoundException(id);
         }
+
+        ClientDAOImpl clientDAO = new ClientDAOImpl();
+        ClientType clientType = clientDAO.getClientByEndpointId(session, endpointType);
+
+        if (clientType == null) {
+            throw new ClientNotFoundException("Client not found for the given endpoint id: " + id);
+        }
+
+//        clientType.getAcl().forEach(acl -> {
+//            if (acl.getEndpoint().getId().equals(id)) {
+//                session.delete(acl);
+//            }
+//        });
         session.delete(endpointType);
-        session.flush();
     }
 
     /**
